dev - Warning errors fixed
## v1.0.0 Jun. 22, 2020 
**Added:** 
- Fix warnings form the project.

+++ buCore/include/buCoreTexture2D.h
@@ -24,8 +24,8 @@ namespace buEngineSDK {
    * @brief virtual Method that initialize the texture object.
    */
   virtual void
-  init(WString filepath, uint32 texFormat, uint32 usage, int32 width,
-       int32 height, int32 mipLevels, int32 arraysize, int32 samplerDescCount,
+  init(WString filepath, uint32 texFormat, uint32 usage, float width,
+       float height, int32 mipLevels, int32 arraysize, int32 samplerDescCount,
        int32 samplerDescQuality, uint32 bindflags, uint32 CPUAccessFlags,
        uint32 miscFlags) = 0;
